backend/requirements.txt
annotated-types==0.7.0
anyio==4.5.2
appdirs==1.4.4
asgiref==3.8.1
attr==0.3.1
attrs==24.2.0
azure-core==1.32.0
azure-storage-blob==12.24.0
bleach==5.0.1
boto==2.49.0
boto3==1.35.72
botocore==1.35.72
boxing==0.1.4
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
cryptography==44.0.0
defusedxml==0.7.1
Deprecated==1.2.15
distro==1.9.0
Django==3.2.25
django-annoying==0.10.6
django-cors-headers==3.6.0
django-csp==3.7
django-debug-toolbar==3.2.1
django-environ==0.10.0
django-extensions==3.1.0
django-filter==2.4.0
django-model-utils==4.1.1
django-ranged-fileresponse==0.1.2
django-rq==2.5.1
django-storages==1.12.3
django-user-agents==0.4.0
djangorestframework==3.13.1
drf-dynamic-fields==0.3.0
drf-flex-fields==0.9.5
drf-generators==0.3.0
exceptiongroup==1.2.2
expiringdict==1.2.2
google-api-core==2.23.0
google-auth==2.36.0
google-cloud-appengine-logging==1.5.0
google-cloud-audit-log==0.3.0
google-cloud-core==2.4.1
google-cloud-logging==3.11.3
google-cloud-storage==2.18.2
google-crc32c==1.5.0
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
grpc-google-iam-v1==0.13.1
grpcio==1.68.1
grpcio-status==1.68.1
h11==0.14.0
htmlmin==0.1.12
httpcore==1.0.7
httpx==0.28.0
humansignal-drf-yasg==1.21.10.post1
idna==3.10
ijson==3.3.0
importlib_metadata==8.5.0
inflection==0.5.1
isodate==0.7.2
jiter==0.8.0
jmespath==1.0.1
joblib==1.4.2
jsonschema==3.2.0
label-studio==1.13.1
label-studio-sdk==1.0.5
launchdarkly-server-sdk==8.2.1
lockfile==0.12.2
lxml==5.3.0
nltk==3.9.1
numpy==1.24.4
openai==1.55.3
opentelemetry-api==1.28.2
ordered-set==4.0.2
packaging==24.2
pandas==2.0.3
pillow==10.4.0
proto-plus==1.25.0
protobuf==5.29.0
psycopg2-binary==2.9.9
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.2
pydantic_core==2.27.1
pyRFC3339==2.0.1
pyrsistent==0.20.0
python-dateutil==2.9.0.post0
python-json-logger==2.0.4
pytz==2022.7.1
PyYAML==6.0.2
redis==3.5.3
regex==2024.11.6
requests==2.32.3
requests-mock==1.12.1
rq==1.10.1
rsa==4.9
rules==2.2
s3transfer==0.10.4
semver==3.0.2
sentry-sdk==2.19.0
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.2
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2024.2
ua-parser==0.18.0
ujson==5.10.0
uritemplate==4.1.1
urllib3==1.26.20
user-agents==2.2.0
vboxapi==1.0
webencodings==0.5.1
wrapt==1.17.0
xmljson==0.2.1
zipp==3.20.2
